requirements.txt
streamlit
torch
transformers
PyPDF2
langchain
langchain-community
python-dotenv
faiss-cpu
sentence-transformers
